Refactor API layer: move interfaces to controllers, update AGENTS.md
